create an activity for the full upsert
